Adding Source Code, Makefile, and Updating README.md:
The source code for the project is included, with the
implementation to demonstrate the use of hamming
codes to correct a randomized error in a byte array.

I include a Makefile, but this is a one-file project
this it is really just for convenience and does not
hide any sophisticated build mechanisms. Perhaps if I
expand on this project in the future it will be more
necessary to have.

The README.md is also updated with a description of
this project.

Signed-off-by: mrandronov <mrandronov88@gmail.com>